itsModel with inheritance approach
Using inheritance to reduce code replicated in different places.
